Remove duplicated code, plus some more cleanup
