+++ src/tracks/battle_graph.hpp
@@ -23,21 +23,13 @@
 #include <string>
 #include <set>
 
-#include <dimension2d.h>
+#include "tracks/graph_structure.hpp"
 #include "tracks/navmesh.hpp"
 
-namespace irr
-{
-    namespace scene { class ISceneNode; class IMesh; class IMeshBuffer; }
-    namespace video { class ITexture; }
-}
-using namespace irr;
-
-class FrameBuffer;
+class GraphStructure;
 class Item;
 class ItemManager;
 class Navmesh;
-class RTT;
 
 /**
 * \ingroup tracks
@@ -49,30 +41,16 @@ class RTT;
 *    design pattern to create an instance).
 \ingroup tracks
 */
-class BattleGraph
+class BattleGraph : public GraphStructure
 {
 
 private:
     static BattleGraph        *m_battle_graph;
 
-    RTT* m_new_rtt;
-
     /** The actual graph data structure, it is an adjacency matrix */
     std::vector< std::vector< float > > m_distance_matrix;
     /** The matrix that is used to store computed shortest paths */
     std::vector< std::vector< int > > m_parent_poly;
-     /** For debug mode only: the node of the debug mesh. */
-    scene::ISceneNode       *m_node;
-    /** For debug only: the mesh of the debug mesh. */
-    scene::IMesh            *m_mesh;
-    /** For debug only: the actual mesh buffer storing the quads. */
-    scene::IMeshBuffer      *m_mesh_buffer;
-
-    /** The minimum coordinates of the quad graph. */
-    Vec3                     m_min_coord;
-
-    /** Scaling for mini map. */
-    float                    m_scaling;
 
     /** Stores the name of the file containing the NavMesh data */
     std::string              m_navmesh_file;
@@ -81,8 +59,6 @@ private:
 
     void buildGraph(NavMesh*);
     void computeFloydWarshall();
-    void createMesh(bool enable_transparency=false,
-                    const video::SColor *track_color=NULL);
 
     BattleGraph(const std::string &navmesh_file_name);
     ~BattleGraph(void);
@@ -115,7 +91,8 @@ public:
     // ----------------------------------------------------------------------
     /** Returns the number of nodes in the BattleGraph (equal to the number of
     *    polygons in the NavMesh */
-    unsigned int      getNumNodes() const { return m_distance_matrix.size(); }
+    virtual const unsigned int getNumNodes() const
+                                        { return m_distance_matrix.size(); }
 
     // ----------------------------------------------------------------------
     /** Returns the NavPoly corresponding to the i-th node of the BattleGraph */
@@ -131,15 +108,13 @@ public:
     const std::vector < std::pair<const Item*, int> >& getItemList()
                                         { return m_items_on_graph; }
 
-    void              createDebugMesh();
-    void              cleanupDebugMesh();
     void              findItemsOnGraphNodes();
-    void              makeMiniMap(const core::dimension2du &where,
-                                  const std::string &name,
-                                  const video::SColor &fill_color,
-                                  video::ITexture** oldRttMinimap,
-                                  FrameBuffer** newRttMinimap);
-    void              mapPoint2MiniMap(const Vec3 &xyz, Vec3 *out) const;
+    // ------------------------------------------------------------------------
+    /** Sets the type of this graph. */
+    virtual void      setType() { m_graph_type = GraphType::GT_BATTLE; }
+    // ------------------------------------------------------------------------
+    virtual const std::vector<GraphNode*> getAllNodes() const
+                                {return std::vector<GraphNode*>();}
 };    //BattleGraph
 
 #endif
